requirements.txt
opencv-contrib-python>=4.7.0.72
pi-sht1x>=1.0.10
RPi.GPIO>=0.7.1
spidev>=3.6
requests